refactor(admin): split quantercise health panels
move the Quantercise health route display panels, skeleton, and server data loader into a sibling section module so the route file stays focused on layout and suspense.

verify: pnpm exec prettier --check apps/admin/src/app/quantercise/health/page.tsx apps/admin/src/app/quantercise/health/health-sections.tsx

verify: git diff --check -- apps/admin/src/app/quantercise/health/page.tsx apps/admin/src/app/quantercise/health/health-sections.tsx

verify: pnpm --filter @anipotts/admin typecheck

verify: pnpm --filter @anipotts/admin lint

verify: pnpm --filter @anipotts/admin build

verify: pnpm validate
